Downloads from Google Docs

Download edited version of a file from Google Drive updating the local version with the new changes.

Use with caution as local version of the file will be overwritten!

To know more about trackdown workflow and features see trackdown-package help page.

download_file(
  file,
  gfile = NULL,
  gpath = "trackdown",
  shared_drive = NULL,
  rm_gcomments = FALSE,
  force = FALSE
)

Arguments

file

character. The path of a local `.Rmd` or `.Rnw` file.

gfile

character. The name of a Google Drive file (defaults to local file name).

gpath

character. (Sub)directory in My Drive or a shared drive (optional). By default files are uploaded in the folder "trackdown". To specify another folder the full path is required (e.g., "trackdown/my_folder"). Use NULL to upload directly at the root level, although it is not recommended.

shared_drive

character. The name of a Google Drive shared drive (optional).

rm_gcomments

[experimental] logical value indicating whether or not to remove Google comments.

force

logical value indicating whether to skip confirm check by user (default is FALSE).

Value

`TRUE` if file from Google Drive was saved, `FALSE` otherwise.